Changes into recovery: - auto creating init.d folder if not present - new option to enable/disable sched_mc from neak menu - small bugfixes

+++ initramfs/sbin/post-init.sh
@@ -69,10 +69,14 @@ echo $(date) START of post-init.sh
 setprop debug.performance.tuning 1; 
 setprop video.accelerate.hw 1;
 setprop debug.sf.hw 1;
-setprop windowsmgr.max_events_per_sec 80;
 
 # Enable SCHED_MC
-echo "1" > /sys/devices/system/cpu/sched_mc_power_savings
+if [ -f /system/etc/schedmc ]; then
+	echo "1" > /sys/devices/system/cpu/sched_mc_power_savings
+else
+	echo "0" > /sys/devices/system/cpu/sched_mc_power_savings
+fi;
+
 # Enable AFTR
 echo "3" > /sys/module/cpuidle/parameters/enable_mask
